+++ docs/CHANGE_LOG.md
@@ -17,6 +17,7 @@
 
 #### **Fixes:**
 - Fixed a fault that caused the dialog moving to sometimes fail when moving the mouse quickly.
+- Fixed documentation stating that events get fired when they don't.
 
 <br>